+++ src/printer.js
@@ -40,8 +40,10 @@ function canAttachComment(node) {
 // where it needs to attach the comments.
 function getCommentChildNodes(node) {
   switch (node.type) {
+    case "heredoc":
+      return [node.beging];
     case "rescue":
-      return (node.body[0] || []).concat(node.body.slice(1));
+      return [].concat(node.body[0]).concat(node.body.slice(1));
     case "aryptn":
       return [node.body[0]]
         .concat(node.body[1])