+++ test/js/alias.test.js
@@ -3,6 +3,15 @@ const { long, ruby } = require("./utils");
 describe("alias", () => {
   test("bare word aliases", () => expect("alias foo bar").toMatchFormat());
 
+  test("bare word operator aliases", () =>
+    expect("alias << push").toMatchFormat());
+
+  test("bare word keyword aliases", () =>
+    expect("alias in within").toMatchFormat());
+
+  test("bare word constant aliases", () =>
+    expect("alias in IN").toMatchFormat());
+
   test("symbol aliases become bare word aliases", () =>
     expect("alias :foo :bar").toChangeFormat("alias foo bar"));