+++ test/js/regexp.test.js
@@ -33,4 +33,10 @@ describe("regexp", () => {
 
     return expect(content).toMatchFormat();
   });
+
+  test("forces braces if could be ambiguous with space", () =>
+    expect("foo %r{ bar}").toMatchFormat());
+
+  test("forces braces if could be ambiguous with equals", () =>
+    expect("foo %r{= bar}").toMatchFormat());
 });